VP side of fixed observable processing
Unfortunately this turned out to be harder than expected.

Basically we need a thing that behaves mostly like a Dataset but not
quite in that it doesn't have fktables in principle, it doesn't have
cuts (but it could) but can have covariance matrices and needs to be
able to be grouped, and so it needs plotting
files as well. The covariance matrix behaves the same as the rest of the
data, so e.g. systematics can be grouped by experiment. The training and
validation masks work the same as for the normal datasets and the
training fraction can be set. We should think what we want to do for the
datasets with one point and such.

The annoying trade offs are that the code pretty much requires to
consider the most general case from the start, unless some undesirable
hacks are applied to make it less functional and pay for them later. For
one I had not considered the grouping originally, so the fixed
observables were repeated in each experiment when passed to n3fit, and
so that piece of code had to become more convoluted. An alternative
would be to hack  the n3fit data thing to take some input that is not
grouped, but that would be slightly incorrect and somewhat gross. I regard much of
this as irreducible complexity, at least insofar I am unable to reduce
it... I considered an alternative approach where Dataset is tricked into
being a fixed observable, but I found it worse.

The way it works is that one has a file in the same format as the
cfactors in a `fixed` folder in the theory format, with the name
`FIXED_<dataset>.dat`, like it is already done in here

https://github.com/LucaMantani/EFTPDF-top/tree/main/theory_predictions/files_for_theory_200/fixed

In addition there must be fitted bsm cfactors, and the usual commondata
files and plotting files. A minimal PLOTTING file could be

```
dataset_label: myexp
experiment: ATLAS
nnpdf31_process: TOP
```

where these three keys are required for the basic functionality. With
that one can do things like

```python
import ruamel_yaml as yaml

from validphys.api import API

inp = yaml.safe_load("""
bsm_fac_data: ["W", "Y"]

theoryid: 200

fixed_observable_inputs:
    - dataset: ATLAS_CMS_WHEL_8TEV

""")

fo = API.fixed_observables(**inp)[0]
```

assuming you invent the requisite BSM factor files.

The implementation in n3it is comparatively easier and will follow
shortly.

There are a number of things that can be improved here:

   - Add tests.
   - Add proper docs on the new functionality.
   - Consider avoiding vandalizing various basic providers like the
   covmat ones and create specific ones instead.